feat(api-lookup-v0): Refactor, add EntityAPILookup
Got rid of some overcomplicated stuff and some premature optimisation. If someone wants to add a cache they can do it themselves, but given you can build an API using dynamic data from Entities, ItemStacks and BlockEntities, it's probably for the better that we don't cache stuff ourselves and let developers figure out when to do it.